Ajout du constante pour type de compte + condition avec type de compte pour l'affichage des boutons

+++ Code/src/view/annonceDetaillee.php
@@ -42,7 +42,8 @@ ob_start();
                     <br>
                     <?php if (!isset($_SESSION['id']) || $_SESSION['id'] !== $annonce['user_id']) :?>
                     <a href="index.php?action=contacterAnnonce&annonceId=<?=$annonce['id']?>" class="btn btn-primary">Contacter</a>
-                    <?php elseif (isset($_SESSION['id']) && $_SESSION['id'] == $annonce['user_id']) :?>
+                    <?php endif;?>
+                    <?php if(isset($_SESSION['id']) && $_SESSION['id'] == $annonce['user_id'] || $_SESSION['userType'] == TYPE_ADMIN) :?>
                         <a href="index.php?action=editAnnonce&annonceId=<?=$annonce['id']?>" class="btn btn-warning">Modifier</a>
                         <a href="index.php?action=deleteAnnonce&annonceId=<?=$annonce['id']?>" class="btn btn-danger">Supprimer</a>
                     <?php endif; ?>
